xcape, xkbOptions
Configure caps lock so it acts as control when used as a modifier key,
and escape when tapped alone (xcape). And configure pressing both shift
keys to toggle caps lock

Unfortunately, Gnome clobbers the xkbOptions value, so I've included it
simply to document the intent (and maybe it will be fixed at some
point). The workaround is easy enough: I've added an executable script
in my dotfiles to set the appropriate options using gsettings.