networking.firewall.allowedTCPPorts
I was trying to send a file over netcat and was blocked by the firewall,
which defaults this option to an empty list. I'm setting it explicitly
so it will be easier to find and modify in the future whenever I want to
open up some ports for incoming traffic.