reconcile session & user xkbOptions
This gets a little complicated under Gnome, because of how gdm cascades
session xkb-options and user xkb-options.

    ┌—— session options ———┐
    │                      │
    │  ┌— user options —┐  │
    │  │                │  │
    └——┴————————————————┴——┘

Gnome's default xkb-options is set to an empty list. Setting it via
gsettings in session commands will affect the graphical login. It will
also affect the logged in user session as long as that user has not set
their own settings via gsettings.

To clean out user settings and allow the session setting to take effect:

  # remove user-level settings
  gsettings reset org.gnome.desktop.input-sources xkb-options

  # That reset to the default empty array. Rebooting allows the
  # session setting to take effect next time we log in.
  reboot