cleaner xkbOptions
It turns out it was not necessary to set this option in sessionCommands.
All that's necessary is to set this option in the nix configuration and
if it's been previously set in either session-commands or a logged in
user shell, pick up the new setting by running:

  gsettings reset org.gnome.desktop.input-sources xkb-options

So I set nix's xkbOptions and did a nixos-rebuild. Then I ran the above
command in a user shell. Then I added the above command to nix's
sessionCommands (maybe I could've just run it in the Gnome console
Alt+F2), did a nixos-rebuild and reboot. Then I removed it from
sessionCommands and did another rebuild & reboot. Wow.