vim colorscheme (coot) plugin
This was easier than I expected. Take something that's part of my local
vim configuration, extract it to its own repo, mirroring the directory
structure it came from, register it as a plugin with VAM, it gets added
to the runtimepath and is available system-wide.

Next step is to extract some of the stuff in my autoload directory.