vim: mline, pinnacle, fugitive
I'm not yet sure how to express the dependency mline has on fugitive and
pinnacle. I tried using an addon-info.json file in mline, but that
breaks because pinnacle is not in any package index that VAM is aware
of. For now, this flat structure works, though it would break if either
fugitive or pinnacle were removed...