vim-nix
This provides better support for writing nix expressions in vim.
Apparently I may also want to set

   nixpkgs.config.vim.ftNix = false;

to prevent the more minimal nix-support from being patched into vim,
because it can clobber some of the behavior of the vim-nix plugin, but
I'm going to avoid doing this until I see the actual problems it is
supposed to address.