tmux: give my pinky a rest (M-Space prefix)
Provide a way to switch the tmux prefix on the fly. It still starts out
at C-a, but when I want to give my left pinky a rest, C-a M-Space will
switch the prefix to M-Space. From there, I can hit M-Space C-a to
switch back to the original setting.

The switch-back is a bit of a hack: it simply re-sources the whole tmux
conf. I don't know how else this could be accomplished given the limited
functionality provided by the tmux configuration language.