overlays: take 1
Translate my package-overrides to an overlay. When I first figured out
how to write a package-override, I wondered if there was a better way to
accomplish what I was doing, since I wasn't really overriding anything,
but rather was adding additional packages outside the nixpkgs tree.

The first road-bump I ran into with this was in the kernel module
derivation, which had bound self and super variables already. I figured
out which ones to use for grabbing stdenv, fetchFromGitHub, and kernel
through trial and error, and I would like to better understand what's
going on here.

Initially I was shadowing the outer "self" and "super" variables
with identically-named variables in the function that extended the
linux kernel, and this cause the rebuild to fail because there is no
fetchFromGitHub attribute on the inner self. Why does the inner self
have a "kernel" attribute but no "fetchFromGitHub" attribute? I also
found I could omit the "stdenv" argument in this function call, but not
the other two. Why?

As a next step, I'd like to split these into separate overlays, probably
defining them in the files that currently define the derivations.