overlay: super.callPackage infers attrs from self
The way callPackage is defined, it will implicitly pass the required
arguments to a package derivation pulling them from the "self" scope,
even though we're calling super.callPackage. This lets me omit the
explicit args.

I also sorted out the issues I was having with the kernel module in this
overlay. As long as that inner functon uses it's inner "super" argument
to run callPackage, everything works as intended. I renamed the inner
args self' and super' to avoid shadowing the outer args (that shadowing
prevented the "kernel" attribute from being inferred, breaking the
build).