vim: turn off gui support
I don't think I get any use out of gui support (I'll find out if that's
not the case), and I've experience a disappearing-cursor bug when using
the Grepper plugin that only occurs if Vim was compiled with gui
support. https://github.com/mhinz/vim-grepper/pull/156

I tried playing with other compilation flags but could not find anything
else that contributed to the bug. It would be nice to know exactly what
causes the disappearing cursor, but at this point I'm more interested in
avoiding it and moving on. If it turns out I actually want gui support,
I'll turn it back on and use a patched version of Grepper, since I found
a benign hack that fixes the cursor redraw issue. I'm planning on
writing my own plugin to use instead of Grepper anyway, which will also
allow me to apply such a fix freely.