vim: re-enable gui support
I had previously disabled gui support because I was experiencing a
disappearing cursor when a Grepper search failed. After upgrading to
NixOS 18.03, Gnome 3.28, and the latest version of Vim, it seems this
is no longer an issue. I suspect it was the Gnome upgrade that actually
fixed it, but that's little more than a guess.
https://github.com/mhinz/vim-grepper/pull/156