vim.gui = "no"
I'm turning this off again because I'm experiencing more cursor redraw
issues with the vim-grepper plugin. I'll run a search, cursor jumps to
the quickfix list results, as expected. Hit Enter to open a result in
the main window, still normal behavior. Run another search and the
cursor appears to remain in the main window, even after the quickfix
list has been populated with new search results. Hit a movement key
(e.g. "j"), and the redraw makes it clear that the cursor was actually
in the quickfix list but was never redrawn there.