overlays: fzf.overrideAttrs (newer version)
This is total overkill in order to pick up a recent fix in fzf.
  https://github.com/junegunn/fzf/issues/1268

I'm committing it mostly in order to demonstrate how overrideAttrs can
be used in an overlay, in this case to pick up a more recent version.
The fix this particular update picks up is not worth the overhead of
maintaining another overlay, so I'll actually be reverting this commit.
I can live without this particular fix until a new release of fzf makes
it into nixpkgs.