openvpn, easyrsa
Set up openvpn with a client config. In the process of setting this up,
I followed some tutorials on setting up an openvpn server, and used
easyrsa to set up the public key infrastructure. I probably won't be
doing that in practice, but I'm keeping easyrsa around for reference,
since it's well-written and thoroughly commented.

The certs and keys that this client config relies on were created as
follows:

  DIR=$(mktemp --tmpdir --directory openvpn-sumall.XXXXXXXXXX)

  USERNAME=ibrennan
  CRTS=( $USERNAME.crt ca.crt )
  KEYS=( $USERNAME.key ta.key )
  PASS=$USERNAME.pass

  for f in ${CRTS[@]} ${KEYS[@]}; do
    lpass show "SumAll OpenVPN ${f}" --format='%fv' \
      | sed -e '1,2 d' -e '${/^$/d;}' \
      > ${DIR}/${f}
  done

  lpass show "SumAll OpenVPN ${USERNAME} passphrase" \
    | awk '/^Notes: /{ print $2 }' \
    > ${DIR}/${PASS}

  sudo bash -c "
    mkdir -p /etc/nixos/services/openvpn/sumall
    for g in ${CRTS[*]} ${KEYS[*]} ${PASS}; do
      cp ${DIR}/\${g} /etc/nixos/services/openvpn/sumall/
    done
    for g in ${KEYS[*]} ${PASS}; do
      chmod 600 /etc/nixos/services/openvpn/sumall/\${g}
    done
  "

  rm -rf $DIR
  unset DIR USERNAME CRTS KEYS f

The updateResolvConf option adds the following to the client config:
  script-security 2
  up ''${pkgs.update-resolv-conf}/libexec/openvpn/update-resolv-conf
  down ''${pkgs.update-resolv-conf}/libexec/openvpn/update-resolv-conf

In order to update /etc/resolv.conf, openvpn needs elevated privileges.
As a result, the following settings will cause failures, since they
downgrade privileges once openvpn has initialized:
  user nobody
  group nogroup

The symptom I observed was that upon stopping the client, dns stopped
working entirely (in a web browser, for example). The contents of
/etc/resolv.conf were those that the vpn server had pushed to the
client, rather than what they should have been restored to upon stopping
the vpn service (1.1.1.1 and so on).

It sounds like downgrading privileges can also cause a client reconnect
failure if the server initiates a restart.

An interesting solution configures openvpn to run as an unprivileged
user without these problems:
  https://community.openvpn.net/openvpn/wiki/UnprivilegedUser

It's well-detailed, but I haven't tried it yet. I suspect it will
take a bit of work to implement it within the confines of NixOS. It
also mentions that it may not work on Ubuntu. The most frequently
suggested solution I see on the openvpn support forums is to just remove
user/group "nobody" settings from your config, so that's what I'm doing
for now.