refine man() function
The man command has a bunch of options and varying signatures (most of
which I never use). Fortunately, all of these less common uses can be
recognized by the presence of either more than one argument or an
argument that starts with -

Restrict my man() function's overrides to the vanilla use-case.