sumall-cli: remove fetchgitPrivate dependency
Using fetchgitPrivate to fetch the sumall-cli source code required using
socat to share SSH_AUTH_SOCK with the nixbld group in order for the Nix
builder daemon to authenticate with GitHub. While this was an
interesting solution, I was unable to get it to work on non NixOS
systems. Instead, I'm using the "private" attr of fetchFromGitHub, which
relies on the GitHub username and password being available in env vars.