vln: set executable bit on install
When I first started using the `install` command in my installPhase
scripts, I didn't specify the permission mode, and it defaulted to 755,
which nix reduces to 555. But I hadn't intended to set the executable
bit on most of these files, so I changed all my `install` calls to
specify `-m644` (which nix reduces to 444). I failed to notice that this
sweeping change disabled the executable bit of vln. This fixes that.