services.timesyncd.enable = false;
This is how you disable the systemd NTP client daemon, allowing you to
manually change the time, e.g. with

  timedatectl set-time '2019-01-18 14:00:00'

I don't plan on keeping this setting, but wanted to document it for
future reference, as it's useful when you want to manually set the
clock in order to repro a date/time-dependent bug.