bashrc: ghcWithPackages()
I can never remember this expression when I need it. Wrapping it in a
function will let me call it when my memory is fuzzy (enter "ghc" and let
tab completion take care of the rest) and provide an easy reference when
I want to see how it's done (enter "type ghcWithPackages" to see the
function definition).