gtk-key-theme
There may still be a better way to achieve this, but this is the only
way that worked so far. Though the arch wiki says that system-wide
settings can be placed in
  /etc/gtk-2.0/gktrc
  /etc/gtk-3.0/settings.ini
those locations weren't being picked up, so I'm manually adding
  /etc/xdg
to XDG_CONFIG_DIRS, in which gtk apps will search for settings.

https://github.com/bennofs/etc-nixos/blob/5fe80d256d0101b1e73c9036d0a7ae6cee2c4267/conf/desktop.nix