i3: gnome-keyring-daemon
I don't know why the keyring daemon appeared to be running under i3
before. Enabling it in the nixos configuration causes it to be activated
through dbus with a autostart desktop entry that specifies:

  OnlyShowIn=GNOME;XFCE;Unity;

Now that it's _not_ running in i3, it's clear how I can set it up
to run in i3. It seems like gpg-agent might be a better option
in general, but it's nixos module doesn't handle sudo very well.
https://github.com/NixOS/nixpkgs/issues/57779

https://discourse.nixos.org/t/i3-and-gnome-keyring/5719/4?u=ivanbrennan