space2meta
This is probably not a keeper. The way it interfers with normal usage of
the space key makes things feel weird. A clear example is the fact that
if you hold the real super key down and hold down some key X, you get a
series of Super+X events, but if you hold down Space to emulate Super
and hold down key X, you get nothing until the moment X is released, at
which point you get a single X event.