hsetroot -solid "#112026"
Getting the background color/image correctly configured has been a bit
tricky due to the interactions of the display manager, compton, and
NixOS.

NixOS is doing the following. If ~/.background-image exists, it uses feh
to set that as the background, otherwise, it uses xsetroot to set the
root X window's background to black:
  xsetroot -solid black

This happens _after_ displayManager.sessionCommands have been run,
clobbering any color that was previously set. A fix for this has been
merged upstream: https://github.com/NixOS/nixpkgs/pull/78346

Also relevant is a newly introduced option:
  services.xserver.windowManager.i3.extraSessionCommands
https://github.com/NixOS/nixpkgs/pull/49492#issuecomment-579311286

In the meantime, my workaround has been to create an image with the
color I want and copy it to ~/.background-image.

While trying to figure out how to get conky to work in i3, I learned
that compton doesn't play nice with xsetroot. I haven't dug into too
much detail, but it behaves as if compton introduces another layer of
background that hides the layer xsetroot touches. Instead, using
hsetroot has the desired affect.