unstable: 39247f8d04c04b3ee629a1f85aeedd582bf41cac
These are the changes I made after updating to the above revision of the
nixos-unstable channel. As described in:
  https://discourse.nixos.org/t/luks-crypto-module-aes-x86-64-not-found-in-linux-kernel-5-4-x/6616
this showed a fatal error (though it didn't fail the build) related to
luks and the aes_x86_64 module not being present in the linux kernel.
That module was removed in kernel 5.4.1
  https://bugs.archlinux.org/task/64721

Rather than reboot and pray (and potentially get locked out of the
system if it's unable to decrypt the drive), I rolled back the channel
update:
  nix-channel --rollback

I'm going to revert this commit and rebuild.