alacritty: update Vi Mode keys
Use Shift-Space to enter Vi Mode, but not to exit it (exit using either
Q or Escape). This frees up Shift-Space within Vi Mode for scrolling up
a page. That's particularly nice in that it allows me to tap Shift-Space
twice to enter Vi Mode and scroll up one page. As a corollary, bind
Space in Vi Mode to scroll down a page.