xmonad: RefocusLast, namedScratchpadAction
The RefocusLast module is exactly what I was looking for, and it can be
used to fix the behavior of NamedScratchpad so that focus returns to the
previously focused window when you banish the scratchpad. I had to
reimplement someNamedScratchpadAction to apply the refocus logic.
There's probably a cleaner way to achieve this if we make
NamedScratchpad more composable.