alacritty-unwrapped
Provide easier access to the underlying executable. This is useful
because alacritty won't let an invocation specify --config-file more
than once, which prevents an invocation of the wrapped executable from
specifying a different config file than the one I've baked into the
wrapper. Specifying a different config file is useful when trying out
config options, as it avoids the need to run a nixos-rebuild.