vim-vue
Looks like vim now comes with a syntax file that eliminates my need for
this plugin, so I won't be keeping it around. I wanted to commit it
though because I ran into difficulty getting it to build (it has a
Makefile that runs tests and assumes the build environment has vim and
git installed), and I want to document how I was able to deal with that.