systemd.tmpfiles: /run/openvpn (status files)
Demonstrate how to provide a service with a place to put volatile files,
and how to configure openvpn to write status to a file regularly. I'm
looking for a programmatic way to check which (if any) openvpn clients
are running, in order to provide a dmenu-powered interface for
starting/stopping clients. This didn't turn out to be a solution to that
problem, because the status files are left in place after the client is
stopped (meaning I can't simply equate the presence of a status file
with an active status for the corresponding client), but I wanted to
document how systemd.tmpfiles rules can be used.