udisks, gvfs, thunar-volman
Provide access to automounted drives (e.g. USB flash drives), as well as
trash support in thunar.

With the services.udisks2 enabled, a systemd udisks2.service is
provided. It will start on demand via dbus, for example, by plugging a
flash drive in and running udisksctl (note that simply plugging the
drive in is not sufficient to start the service).

Similarly, with gvfs and thunar-volman enabled, the service will be
started if you plug in a flash drive and run thunar.

The systemd service runs udisksd. Even with this running, the drive will
not be auto-mounted. It can be mounted via `udisksctl mount ...` or by
attempting to access the drive in thunar.