dmenu_cdpath: double fork
The shell process that was piping dmenu output to xargs in order to
invoke alacritty was waiting for alacritty's exit code. If we instead
double fork, we avoid keeping that process around unnecessarily. The
xargs pipeline runs in a subshell and alacritty runs as a background
child of the subshell. The subshell won't wait for the background
process, so it immediately terminates, and the alacritty process gets
reparented to...

Actually, this isn't quite right yet. While it avoids keeping the full
dmenu-xargs pipeline, it still keeps xargs hanging around.