dmenu_manpages
Provide a dmenu-based alternative to the manpage function I was relying
on an XMonad.Prompt for. Unfortunately, XMonad.Prompt is slow, causing a
half second delay between the time I hit the key and the time the prompt
appears. Also unfortunate is the fact that XMonad submaps are
incompatible with dmenu, since the submap blocks waiting for keyboard
input.