kmonad: more adjustment
I'd rather err on the side of treating / as a / rather than as a Shift.
When I want the key to be treated as a Shift modifier, I'm generally a
little more deliberate, and hold it down a bit longer, so let's try
relying on the tap-hold timeout to make that distinction.