feat(cachix): cachix use nix-community
I want to start using this to facilitate pulling cached builds for the
emacs overlay.

I was hesitant to perform the sort of imperative installation that
cachix recommends, but after reading through the source code, I was
convinced to pull the trigger.

It looks like there is also a more declarative approach, but it's not
clear how well-supported it is:
https://github.com/cachix/cachix/issues/239
https://github.com/jonascarpay/declarative-cachix

This was performed using the "cachix-1.3.3" release of cachix:
https://github.com/cachix/cachix/releases/tag/v1.3.3

Following are the details of what directories/files were affected by the
process:
- no changes to /etc/nix/nix.conf
- no changes to /etc/nix/netrc
- created /etc/nixos/cachix.nix
- created /etc/nixos/cachix/
- created /etc/nixos/cachix/nix-community.nix

The actual command and its output:
```
$ cachix use nix-community
Cachix configuration written to /etc/nixos/cachix.nix.
Binary cache nix-community configuration written to /etc/nixos/cachix/nix-community.nix.

To start using cachix add the following to your /etc/nixos/configuration.nix:

    imports = [ ./cachix.nix ];

Then run:

    $ sudo nixos-rebuild switch

```

After adding `./cachix.nix` to my imports in configuration.nix I ran a
rebuild.