xmonad: experimental WorkScope
Starting small, this adds the ability to attach a string to a workspace,
change it per workspace, and display the current workspace's associated
string, if present, either in a debug log (via a trace call whose output
is viewable in `journalctl --user -t xmonad`) or in the status bar.

I'm thinking about how to implement something a bit like tmux sessions
for xmonad, where workspaces can be associated with a "scope", and we
can choose a current scope to constrain workspace navigation/display to
workspaces in that scope.

I imagine eventually assigning a workspace a list of scopes in which
it's viewable.

I'd also like to be able to easily detach/reattach a workspace from the
current scope, probably via a dmenu prompt or similar interface.