xmonad: shiftToLayer
I'm not sure the ideal workflow is for this operation, e.g. should the
first number key indicate the destination layer, then a second number
key (in a submap) indicates which workspace within that layer to choose
as the destination?

For now, choosing the most recently focused workspace on the selected
layer seems reasonable.