update users module
Note that in the following commit, I updated the hashed password to use
the latest hashing algorithm, but I only did so for the users module,
which was not actually in use at the time.

  AuthorDate: Fri Dec 23 10:59:52 2022 -0500
  4280236c5e3a005366dc7aba5bbd6bd523266188
  update hashedPassword

As a result, I was still relying on the old hashing algorithm that had
been used in this older commit.

  AuthorDate: Sat Jan 15 23:15:07 2022 -0500
  3da27bd27ad529e343830e984d0d65a894805a02
  thinkpad x1 carbon gen 9: initial config