sops-nix example
I'm not yet sure whether I'd be better served by sops-nix or agenix.
This is just a proof-of-concept. I followed the instructions found at:
https://github.com/Mic92/sops-nix#usage-example

I generated an age key for my user:

    mkdir -p ~/.config/sops/age
    age-keygen -o ~/.config/sops/age/keys.txt

as well as an age key for thinkpad9:

    nix-shell -p ssh-to-age --run 'ssh-to-age -i /etc/ssh/ssh_host_ed25519_key.pub'

Then, as the root user, I set up the secrets configuration to permit
decryption with these keys:

    sudo -i
    cd /etc/nixos/
    mkdir -p sops-nix/secrets
    nix-shell -p sops --run "SOPS_AGE_KEY_FILE=/home/${SUDO_USER:-$USER}/.config/sops/age/keys.txt sops sops-nix/secrets/secrets.yaml"

After a `nixos-rebuild switch`, the example secret is stored in
cleartext with restricted permissions:

    # cat /run/secrets/example_key
    example_value

    # ls -l /run/secrets/example_key
    -r-------- 1 root root 13 Aug 11 23:10 /run/secrets/example_key