agenix example
I'm not yet sure whether I'd be better served by sops-nix or agenix.
This is just a proof-of-concept. I followed the instructions found at:
https://github.com/ryantm/agenix#tutorial

I generated an SSH specifically for use with agenix (see sysadmin-notes
repo for info about mangaging such keys). To generate the key, I ran the
following as a normal (non-root) user:

    cd "${XDG_DATA_HOME:-$HOME/.local/share}"
    mkdir -p encryption-keys/ssh
    ssh-keygen -t ed25519 -N "" -C "ivan@agenix" -f encryption-keys/ssh/id_ed25519_agenix

I then added the public key to agenix/secrets/secrets.nix

    cat "${XDG_DATA_HOME:-$HOME/.local/share}/encryption-keys/ssh/id_ed25519_agenix.pub"

and did the same for the thinkpad9 host public key:

    cat /etc/ssh/ssh_host_ed25519_key.pub

Then, as the root user, I created a secret:

    sudo -i
    cd /etc/nixos/agenix/secrets/
    nix run github:ryantm/agenix -- -e secret1.age -i "/home/${SUDO_USER}/.local/share/encryption-keys/ssh/id_ed25519_agenix"

After a `nixos-rebuild switch`, the example secret is stored in
cleartext with restricted permissions:

    # cat /run/agenix/secret1
    value1

    # ls -l /run/agenix/secret1
    -r-------- 1 root root 7 Aug 12 17:38 /run/agenix/secret1