fix(xmonad|alacritty): copy/paste keys
The previous approach (using sendKey) no longer works, because alacritty
ignores X11 send_events.
https://github.com/alacritty/alacritty/issues/7893

In order to simulate the desired keys using xdotool, we need to first
simulate a keyup of the key we actually pressed (c or v), then simulate
keystroke of the key we've bound in alacritty.toml (Copy or Paste).
https://github.com/jordansissel/xdotool/issues/210