fix pass-otp completions
Bash loads completions from $XDG_DATA_DIRS/bash-completion/completions/
on demand, and the completion function for `pass` is capable of
completing an extension's commands only if the extension's completion
function has already been loaded and registered. The completion script
for pass-otp was never being loaded, since Bash automatically discovers
completions for the command being entered, not for subcommands provided
by an extension.

To fix, we load this particular completion script upfront, so that will
be visible when we're trying to complete a `pass` command.