stop using resound
The resound utility is something I put together to work around an audio
issue that affected older kernel versions. It's no longer necessary.
I've opted to keep the package definition itself, despite no longer
including it in system packages, since it may serve as a useful
reference in the future.