udev rules for trackpoint keyboard
Add udev rules to turn off FnLock when connecting an external Trackpoint
keyboard over either bluetooth or USB wireless.

The keyboard's HID_ID when connected over bluetooth is
"0005:000017EF:000060E1", where 0005 denotes bluetooth, 000017EF denotes
the vendor (Lenovo), and 000060E1 denotes the product.

Similarly, when connected over the USB wireless device, the HID_ID is
"0003:000017EF:000060EE", where 0003 denotes USB, 000017EF denotes the
vendor, and 000060EE denotes the product.

Toggling FnLock by pressing Fn+Esc works when connected over bluetooth,
but not USB. I rarely if ever want to turn it on (I mostly want the
media keys, and am willing to use Fn as a modifier when I actually need
to type a function key), so that's not a problem as long as the keyboard
defaults FnLock to off. In the rare case where I actually did want to
toggle it, I could resort to running:

    echo 1 | sudo tee /sys/bus/hid/devices/0003:17EF:60EE.*/fn_lock

To reduce some of the friction there, I'm using these udev rules to also
make the fn_lock file writeable for members of the "input" group, so the
above command simplifies to:

    echo 1 > /sys/bus/hid/devices/0003:17EF:60EE.*/fn_lock

It would be easy to create a shell script that toggles FnLock, and bind
it to a key. I don't know that it's warranted, since I so rarely want to
turn FnLock on, but for reference, the following script would do the
trick:

    fn_locks=()

    for fn_lock in /sys/bus/hid/devices/0003:17EF:60EE.*/fn_lock
    do
        if [ -f "$fn_lock" ] && [ -w "$fn_lock" ]
        then
            fn_locks+=( "$fn_lock" )
        fi
    done

    if (( ${#fn_locks[@]} == 1 ))
    then
        file=${fn_locks[0]}
        value=$(< "$file")
        if [[ "$value" == "0" ]]
        then
            echo 1 > "$file"
        elif [[ "$value" == "1" ]]
        then
            echo 0 > "$file"
        fi
    fi