fix(interception-tools): device selector
Contrary to what I thought in commit 358763b, removing the LINK from
device selection logic does in fact break the trackpoint pointer on the
external Trackpoint II keyboard.

Fix that by adding LINK back to the existing job. That re-introduces the
original problem: the Keychron keyboard doesn't get caps2esc applied if
connected via Bluetooth, since it has no symlink to match on. To fix
that, add a separate job that looks specifically for bluetooth keyboards
(BUSTYPE 5) with a name suggesting they are a Keychron keyboard.