update/fix bluetooth trackpoint fn_lock
Since I updated the system (cbe6de9d275d2cffbb14756490b372acb633f9c9),
this piece of configuration stopped working. The fn_lock attribute was
defaulting to on ("1") whenever the keyboard was connected. It appears
to be a timing issue, with the fn_lock file not existing until a
fraction of a second after the rule fires. I verified this by
temporarily adding a rule like:

    ACTION=="add", SUBSYSTEM=="hid", ENV{HID_ID}=="0005:000017EF:000060E1", RUN+="${pkgs.runtimeShell} -c 'ls %S%p >> /home/ivan/udev-debug.txt; sleep 1; ls %S%p >> /home/ivan/udev-debug.txt'"