another attempted fix for fn_lock
This works by waiting for a "bind" event to arrive, at which point the
fn_lock sysfs attribute will necessarily be present. It's problematic,
however, because although the kernel appears to report a "bind" event
early on, there's often a long delay (~25s) before udev processes that
event, as far as I can tell.