improve fix for fn_lock
Relying on the "bind" event introduced a long delay before turning off
fn_lock. This commit goes back to matching on the "add" event, and
rather than making the udev rule's RUN command wait synchronously for
fn_lock to appear, it just starts a oneshot systemd service to poll
asynchronously for fn_lock, turn it off once it appears, then exit.

I also tried replacing the udev rule with a systemd path unit,

    systemd.paths.fnlock-off = {
      description = "Disable FnLock when attribute appears for TrackPoint Keyboard II (Bluetooth)";
      wantedBy = [ "multi-user.target" ];
      pathConfig = {
        PathExistsGlob = "/sys/bus/hid/devices/0005:17EF:60E1.*/fn_lock";
      };
    };

That approach is problematic, because it attempts to restart the
fnlock-off service over and over as long as the fn_lock path exists,
unless you set `RemainAfterExit=true` in the service, in which case the
whole mechanism only works once per boot.